backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
google-generativeai>=0.3.0
pillow>=10.3.0
python-dotenv>=1.0.0
python-docx>=1.1.0
python-pptx>=0.6.23
PyPDF2>=3.0.1
reportlab>=4.0.0
openpyxl>=3.1.2